Update dateConversion.py
numpy causes an error in executable and was not necessary

+++ dateConversion.py
@@ -1,5 +1,4 @@
 import datetime as dt
-import numpy
 
 #day1 is a datetime object that is being counted up from in the binary
 day1 = dt.date(1902, 9, 20)
@@ -11,7 +10,7 @@ yesterdayDate = dt.date(yesterday.year, yesterday.month, yesterday.day)
 #convert it to a usuable hex figure, hex1 and 2 from each item in binary are
 #swapped and we need to replicate that
 yesterdaySince = yesterdayDate - day1
-intYesterdaySince = numpy.int64(yesterdaySince.days)
+intYesterdaySince = yesterdaySince.days
 hexYesterdaySince = hex(intYesterdaySince)
 mdYesterdaySince = hexYesterdaySince[4:] + hexYesterdaySince[2:4]
 
@@ -25,10 +24,11 @@ with open('C:\\bclogs\\reportDates.txt', 'r') as f:
         useline = line.split(',')
         rDate = dt.date(int(useline[0]), int(useline[1]), int(useline[2].rstrip('\n')))
         rSince = rDate - day1
-        rSinceInt = numpy.int64(rSince.days)
+        rSinceInt = rSince.days
         rSinceHex = hex(rSinceInt)
         rSinceMD = rSinceHex[4:] + rSinceHex[2:4]
         mdHexSince = '{}c3a82faa'.format(rSinceMD)
+
 
         with open('C:\\bclogs\\file012.dat', 'rb') as f12:
             f12Binary = f12.read()
@@ -38,3 +38,6 @@ with open('C:\\bclogs\\reportDates.txt', 'r') as f:
             path = 'C:\\bclogs\\dateReportConvert\\file012_NEW\\file012_{}.dat'.format(str(rDate))
             with open(path, 'wb') as new_f:
                 new_f.write(new_f12Binary)
+
+
+#save as file_01242016.dat
